pyproject.toml
[tool.poetry]
name = "Karman"
version = "0.1.0"
description = "A management software for UltraStar karaoke files."
authors = ["Kim Wittenburg <codello@wittenburg.kim>"]
license = "MIT"

readme = "README.md"
repository = "https://github.com/Karaoke-Manager/backend"
# homepage = "..."
# documentation = "..."

keywords = ["fastapi", "asyncio", "ultrastar", "karaoke"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Framework :: AsyncIO"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Karaoke-Manager/backend/issues"


[tool.poe.tasks]
serve = { cmd = "uvicorn karman:app --reload --log-config=logging.yml", help = "Runs a development instance" }
test = { cmd = "pytest -n auto --doctest-modules", help = "Run Tests" }
clean = { cmd = "rm -rf .mypy_cache .pytest_cache htmlcov coverage.xml .coverage", help = "Remove build and cache files" }
[tool.poe.tasks.lint]
sequence = [
    "flake8", # Runs black and isort as well
    "mypy"
]
default_item_type = "cmd"
help = "Checks the code style on the entire project."
[tool.poe.tasks.makemigration]
cmd = "alembic revision --autogenerate -m '$message'"
args = [{ name = "message", positional = true, required = true, help = "A short name or message for the migration." }]
help = "Create a migration file based on the current database models."
[tool.poe.tasks.migrate]
cmd = "alembic upgrade '$revision'"
args = [{ name = "revision", positional = true, default = "head", help = "The revision to which you want to upgrade." }]
help = "Applies migrations to the database."
[tool.poe.tasks.rollback]
cmd = "alembic downgrade '$revision'"
args = [{ name = "revision", positional = true, default = "-1", help = "The revision to which you want to rollback. By default the last revision is rolled back." }]
help = "Rolls back one or more migrations."

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
files = ["karman", "tests"]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
# Untyped definitions and calls
disallow_untyped_defs = true
# None and Optional handling
no_implicit_optional = true
# Configuring warnings
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
# Miscellaneous
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["jose.*", "pytest", "_pytest", "_pytest.fixtures", "_pytest.monkeypatch"]
ignore_missing_imports = true

[tool.pydantic-mypy]
warn_untyped_fields = true

[tool.pytest.ini_options]
testpaths = ["tests", "karman"]
asyncio_mode = "auto"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
fastapi-pagination = "^0.9.3"
python-multipart = "^0.0.5"
python-dotenv = "^0.20.0"
orjson = "^3.6.8"
PyYAML = "^6.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }

# ORM and Database Drivers
sqlalchemy = { version = "^1.4.36", extras = ["asyncio", "mypy"] }
alembic = "^1.7.7"
asyncpg = { version = "^0.25.0", optional = true }
aiomysql = { version = "^0.1.0", optional = true }
aiosqlite = { version = "^0.17.0", optional = true }

# Uvicorn Optional Dependency
uvicorn = { version = "^0.17.6", optional = true, extras = ["standard"] }

# Pytest Optional Dependency
pytest = { version = "^7.1.1", optional = true }
pytest-asyncio = { version = "^0.18.3", optional = true }
pytest-cov = { version = "^3.0.0", optional = true }
pytest-xdist = { version = "^2.5.0", optional = true }
requests = { version = "^2.27.1", optional = true }

[tool.poetry.dev-dependencies]
poethepoet = "^0.13.1"
uvicorn = { version = "^0.17.6", extras = ["standard"] }
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
flake8-black = "^0.3.2"
flake8-isort = "^4.1.1"

mypy = "^0.950"
types-PyYAML = "^6.0.1"
types-passlib = "^1.7.5"

[tool.poetry.extras]
uvicorn = ["uvicorn"]
postgresql = ["asyncpg"]
postgres = ["asyncpg"]
mysql = ["aiomysql"]
sqlite = ["aiosqlite"]
tests = ["pytest", "pytest-asyncio", "pytest-cov", "pytest-xdist", "requests"]

[tool.poetry.build]
generate-setup-file = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
